Refactor LLM server into reusable class
